+++ src/lib.rs
@@ -87,6 +87,7 @@ mod kv;
 mod kv_client;
 mod pd;
 pub mod raw;
+mod request;
 mod security;
 mod stats;
 pub mod transaction;
