+++ src/mock.rs
@@ -1,25 +1,44 @@
 // Copyright 2019 TiKV Project Authors. Licensed under Apache-2.0.
 
-use std::{sync::Arc, time::Duration};
+use crate::{
+    kv_client::{KvClient, KvConnect, Store},
+    pd::{PdClient, PdRpcClient, Region, RegionId, RetryClient},
+    request::KvRequest,
+    transaction::Timestamp,
+    Config, Error, Key, Result,
+};
 
 use futures::future::ready;
 use futures::future::BoxFuture;
-
-use crate::Error;
-use crate::{
-    kv_client::{KvClient, KvConnect, KvRequest, Store},
-    pd::{PdClient, PdRpcClient, Region, RegionId, RetryClient},
-    transaction::Timestamp,
-    Config, Key, Result,
-};
-
 use grpcio::CallOption;
 use kvproto::metapb;
+use std::{sync::Arc, time::Duration};
+
+pub fn pd_rpc_client() -> PdRpcClient<MockKvConnect, MockCluster> {
+    let config = Config::default();
+    PdRpcClient::new(
+        &config,
+        |_, _| MockKvConnect,
+        |e, sm| {
+            Ok(RetryClient::new_with_cluster(
+                e,
+                sm,
+                config.timeout,
+                MockCluster,
+            ))
+        },
+    )
+    .unwrap()
+}
 
 #[derive(Clone, Eq, PartialEq, Debug)]
 pub struct MockKvClient {
     addr: String,
 }
+
+pub struct MockKvConnect;
+pub struct MockCluster;
+pub struct MockPdClient;
 
 impl KvClient for MockKvClient {
     fn dispatch<T: KvRequest>(
@@ -31,8 +50,6 @@ impl KvClient for MockKvClient {
     }
 }
 
-pub struct MockKvConnect;
-
 impl KvConnect for MockKvConnect {
     type KvClient = MockKvClient;
 
@@ -40,51 +57,6 @@ impl KvConnect for MockKvConnect {
         Ok(MockKvClient {
             addr: address.to_owned(),
         })
-    }
-}
-
-pub struct MockCluster;
-
-pub struct MockPdClient;
-
-impl PdClient for MockPdClient {
-    type KvClient = MockKvClient;
-
-    fn map_region_to_store(
-        self: Arc<Self>,
-        region: Region,
-    ) -> BoxFuture<'static, Result<Store<Self::KvClient>>> {
-        Box::pin(ready(Ok(Store::new(
-            region,
-            MockKvClient {
-                addr: String::new(),
-            },
-            Duration::from_secs(60),
-        ))))
-    }
-
-    fn region_for_key(&self, key: &Key) -> BoxFuture<'static, Result<Region>> {
-        let bytes: &[_] = key.into();
-        let region = if bytes.is_empty() || bytes[0] < 10 {
-            Self::region1()
-        } else {
-            Self::region2()
-        };
-
-        Box::pin(ready(Ok(region)))
-    }
-    fn region_for_id(&self, id: RegionId) -> BoxFuture<'static, Result<Region>> {
-        let result = match id {
-            1 => Ok(Self::region1()),
-            2 => Ok(Self::region2()),
-            _ => Err(Error::region_not_found(id, None)),
-        };
-
-        Box::pin(ready(result))
-    }
-
-    fn get_timestamp(self: Arc<Self>) -> BoxFuture<'static, Result<Timestamp>> {
-        unimplemented!()
     }
 }
 
@@ -116,19 +88,44 @@ impl MockPdClient {
     }
 }
 
-pub fn pd_rpc_client() -> PdRpcClient<MockKvConnect, MockCluster> {
-    let config = Config::default();
-    PdRpcClient::new(
-        &config,
-        |_, _| MockKvConnect,
-        |e, sm| {
-            Ok(RetryClient::new_with_cluster(
-                e,
-                sm,
-                config.timeout,
-                MockCluster,
-            ))
-        },
-    )
-    .unwrap()
+impl PdClient for MockPdClient {
+    type KvClient = MockKvClient;
+
+    fn map_region_to_store(
+        self: Arc<Self>,
+        region: Region,
+    ) -> BoxFuture<'static, Result<Store<Self::KvClient>>> {
+        Box::pin(ready(Ok(Store::new(
+            region,
+            MockKvClient {
+                addr: String::new(),
+            },
+            Duration::from_secs(60),
+        ))))
+    }
+
+    fn region_for_key(&self, key: &Key) -> BoxFuture<'static, Result<Region>> {
+        let bytes: &[_] = key.into();
+        let region = if bytes.is_empty() || bytes[0] < 10 {
+            Self::region1()
+        } else {
+            Self::region2()
+        };
+
+        Box::pin(ready(Ok(region)))
+    }
+
+    fn region_for_id(&self, id: RegionId) -> BoxFuture<'static, Result<Region>> {
+        let result = match id {
+            1 => Ok(Self::region1()),
+            2 => Ok(Self::region2()),
+            _ => Err(Error::region_not_found(id, None)),
+        };
+
+        Box::pin(ready(result))
+    }
+
+    fn get_timestamp(self: Arc<Self>) -> BoxFuture<'static, Result<Timestamp>> {
+        unimplemented!()
+    }
 }
